requirements.txt
certifi==2019.3.9
Django==2.2
django-filter==2.1.0
djangorestframework==3.9.2
Markdown==3.1
pytz==2018.9
sqlparse==0.3.0
wincertstore==0.2
django-heroku==0.3.1
gunicorn==19.9.0